Fixed creation of mockups
