+++ src/sequelize/mockups/offers.ts
@@ -1,0 +1,46 @@
+import db from '../';
+
+export default function create() {
+    const motorcycle = db['motorcycle']
+    motorcycle.findById(1)
+      .then((moto) => {
+        Promise.all([
+          moto.setAccessorie(1),
+          moto.setMake('KTM'),
+          moto.setModel('SX 125'),
+          moto.setModelspec(1),
+          moto.setMotorspec(1),
+          moto.setProtection(1),
+          moto.setService(1),
+        ]).then(() => console.log(moto.dataValues));
+      })
+      .catch(err => console.log('All error', err));
+
+    // motorcycle.findById(2)
+    //   .then((moto) => {
+    //     Promise.all([
+    //       moto.setAccessorie(1),
+    //       moto.setMake('Husaberg'),
+    //       moto.setModel('FE 390'),
+    //       moto.setModelspec(1),
+    //       moto.setMotorspec(1),
+    //       moto.setProtection(1),
+    //       moto.setService(1),
+    //     ]).then(() => console.log(moto.dataValues));
+    //   })
+    //   .catch(err => console.log('All error', err));
+
+    // motorcycle.findById(3)
+    //   .then((moto) => {
+    //     Promise.all([
+    //       moto.setAccessorie(1),
+    //       moto.setMake('KTM'),
+    //       moto.setModel('Freeride 250R'),
+    //       moto.setModelspec(1),
+    //       moto.setMotorspec(1),
+    //       moto.setProtection(1),
+    //       moto.setService(1),
+    //     ]).then(() => console.log(moto.dataValues));
+    //   })
+    //   .catch(err => console.log('All error', err));
+}
