+++ src/sequelize/mockups/user.json
