+++ src/sequelize/models/offer.ts
@@ -1,25 +1,32 @@
-import * as Sequelize from 'sequelize';
+export default function(sequelize, Sequelize) {
 
-const types = [
-  'club',
-  'coach',
-  'guide',
-  'mechanic',
-  'motocross',
-  'motocycle',
-  'photographer',
-  'transport',
-  'garage',
-];
+  const types = [
+    'club',
+    'coach',
+    'guide',
+    'mechanic',
+    'motocross',
+    'motocycle',
+    'photographer',
+    'transport',
+    'garage',
+  ];
 
-export default {
-  title: {
-    allowNull: false,
-    type: Sequelize.STRING,
-  },
-  type: {
-    allowNull: false,
-    type: Sequelize.ENUM,
-    values: types,
-  },
-};
+  return sequelize.define('offer', {
+    title: {
+      allowNull: false,
+      type: Sequelize.STRING,
+    },
+    type: {
+      allowNull: false,
+      type: Sequelize.ENUM,
+      values: types,
+    },
+  }, {
+    classMethods: {
+      associate(db) {
+        this.belongsTo(db.accessorie);
+      }
+    }
+  });
+}
