+++ src/sequelize/models/offerer.ts
@@ -1,12 +1,12 @@
-import * as Sequelize from 'sequelize';
-
-export default {
-  name: {
-    allowNull: false,
-    type: Sequelize.STRING,
-  },
-  picture: {
-    allowNull: false,
-    type: Sequelize.STRING,
-  },
-};
+export default function(sequelize, Sequelize) {
+  return sequelize.define('offerer', {
+    name: {
+      allowNull: false,
+      type: Sequelize.STRING,
+    },
+    picture: {
+      allowNull: false,
+      type: Sequelize.STRING,
+    },
+  });
+}
