+++ src/sequelize/models/tag.ts
@@ -1,8 +1,8 @@
-import * as Sequelize from 'sequelize';
-
-export default {
-  name: {
-    allowNull: false,
-    type: Sequelize.STRING,
-  },
-};
+export default function(sequelize, Sequelize) {
+  return sequelize.define('tag', {
+    name: {
+      allowNull: false,
+      type: Sequelize.STRING,
+    },
+  });
+}
