+++ src/sequelize/models/user.ts
@@ -1,34 +1,34 @@
-import * as Sequelize from 'sequelize';
-
-export default {
-  name: {
-    type: Sequelize.STRING,
-    allowNull: false,
-  },
-  email: {
-    type: Sequelize.STRING,
-    allowNull: false,
-    unique: true,
-    validate: {
-      isEmail: true,
+export default function(sequelize, Sequelize) {
+  return sequelize.define('user', {
+    name: {
+      type: Sequelize.STRING,
+      allowNull: false,
     },
-  },
-  birthday: {
-    type: Sequelize.DATE,
-    allowNull: false,
-  },
-  password: {
-    type: Sequelize.STRING,
-    allowNull: false,
-  },
-  verified: {
-    type: Sequelize.BOOLEAN,
-    defaultValue: false,
-    allowNull: false,
-  },
-  token: {
-    type: Sequelize.UUID,
-    defaultValue: Sequelize.UUIDV1,
-    allowNull: false,
-  },
-};
+    email: {
+      type: Sequelize.STRING,
+      allowNull: false,
+      unique: true,
+      validate: {
+        isEmail: true,
+      },
+    },
+    birthday: {
+      type: Sequelize.DATE,
+      allowNull: false,
+    },
+    password: {
+      type: Sequelize.STRING,
+      allowNull: false,
+    },
+    verified: {
+      type: Sequelize.BOOLEAN,
+      defaultValue: false,
+      allowNull: false,
+    },
+    token: {
+      type: Sequelize.UUID,
+      defaultValue: Sequelize.UUIDV1,
+      allowNull: false,
+    },
+  });
+}
